refactor(@sounisi5011/encrypted-archive): reduce similar blocks in `src/runtimes/node/compress/gzip.ts` and `src/runtimes/node/compress/brotli.ts`
Code Climate reported:
+ Similar blocks of code found in 2 locations. Consider refactoring.
